Rename ifx_fast to generic_fast and remove generic_fast pooling implementation
