+++ tensorflow/lite/micro/kernels/generic_fast/conv/conv_ops_int8.h
@@ -13,12 +13,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-#ifndef TENSORFLOW_LITE_MICRO_KERNELS_IFX_FAST_CONV_CONV_OPS_H_
-#define TENSORFLOW_LITE_MICRO_KERNELS_IFX_FAST_CONV_CONV_OPS_H_
+#ifndef TENSORFLOW_LITE_MICRO_KERNELS_GENERIC_FAST_CONV_CONV_OPS_H_
+#define TENSORFLOW_LITE_MICRO_KERNELS_GENERIC_FAST_CONV_CONV_OPS_H_
 
 #include "tensorflow/lite/c/common.h"
-#include "tensorflow/lite/micro/kernels/ifx_fast/conv/conv_core.h"
-#include "tensorflow/lite/micro/kernels/ifx_fast/conv/conv_op_data.h"
+#include "tensorflow/lite/micro/kernels/generic_fast/conv/conv_core.h"
+#include "tensorflow/lite/micro/kernels/generic_fast/conv/conv_op_data.h"
 
 namespace tflite {
 namespace ops {
@@ -74,4 +74,4 @@ TfLiteStatus EvalConvInt8(TfLiteConvParams* params, OpData* data,
 }  // namespace ops
 }  // namespace tflite
 
-#endif /* TENSORFLOW_LITE_MICRO_KERNELS_IFX_FAST_CONV_CONV_OPS_H_ */
+#endif /* TENSORFLOW_LITE_MICRO_KERNELS_GENERIC_FAST_CONV_CONV_OPS_H_ */
