+++ tensorflow/lite/micro/kernels/generic_fast/depthwise_conv/depthwise_conv_impl.h
@@ -13,44 +13,44 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-#ifndef TENSORFLOW_LITE_MICRO_KERNELS_IFX_FAST_CONV_CONV_IMPL_H_
-#define TENSORFLOW_LITE_MICRO_KERNELS_IFX_FAST_CONV_CONV_IMPL_H_
+#ifndef TENSORFLOW_LITE_MICRO_KERNELS_GENERIC_FAST_DEPTHWISE_CONV_DEPTHWISE_CONV_IMPL_H_
+#define TENSORFLOW_LITE_MICRO_KERNELS_GENERIC_FAST_DEPTHWISE_CONV_DEPTHWISE_CONV_IMPL_H_
 
 #include "tensorflow/lite/c/builtin_op_data.h"
 #include "tensorflow/lite/c/common.h"
 #include "tensorflow/lite/kernels/internal/common.h"
-#include "tensorflow/lite/kernels/internal/quantization_util.h"
-#include "tensorflow/lite/kernels/internal/reference/integer_ops/conv.h"
 #include "tensorflow/lite/kernels/internal/tensor_ctypes.h"
 #include "tensorflow/lite/kernels/kernel_util.h"
 #include "tensorflow/lite/kernels/padding.h"
-#include "tensorflow/lite/micro/kernels/ifx_fast/conv/conv_op_data.h"
+#include "tensorflow/lite/micro/kernels/generic_fast/depthwise_conv/depthwise_conv_op_data.h"
 #include "tensorflow/lite/micro/kernels/kernel_util.h"
 
 namespace tflite {
 namespace ops {
 namespace micro {
-namespace conv {
+namespace depthwise_conv {
 
 constexpr int kInputTensor = 0;
 constexpr int kFilterTensor = 1;
 constexpr int kBiasTensor = 2;
 constexpr int kOutputTensor = 0;
 
-// Conv is quantized along dimension 0:
+// Depthwise conv is quantized along dimension 3:
 // https://www.tensorflow.org/lite/performance/quantization_spec
-constexpr int kConvQuantizedDimension = 0;
+constexpr int kDepthwiseConvQuantizedDimension = 3;
+
+// Size of the cached buffer we'll be using to hold reordered weights.
+constexpr int kReshapedFilterDataSize = 1 * 1024;
 
 /*
  * Init function is called once at the beginning to initialize kernels and
  * allocate memory.
  */
 void* Init(TfLiteContext* context, const char* buffer, size_t length) {
-  void* raw = context->AllocatePersistentBuffer(context, sizeof(OpData));
-  TFLITE_DCHECK(raw != nullptr);
-  OpData* data = reinterpret_cast<OpData*>(raw);
-  *data = {};
-  return raw;
+  TFLITE_DCHECK(context->AllocatePersistentBuffer != nullptr);
+  void* data = context->AllocatePersistentBuffer(context, sizeof(OpData));
+  TFLITE_DCHECK(data != nullptr);
+  return data;
 }
 
 /*
@@ -58,21 +58,19 @@ void* Init(TfLiteContext* context, const char* buffer, size_t length) {
  * and parameters.
  */
 TfLiteStatus CalculateOpData(TfLiteContext* context, TfLiteNode* node,
-                             TfLiteConvParams* params, int width, int height,
-                             int filter_width, int filter_height, int out_width,
-                             int out_height, const TfLiteType data_type,
-                             OpData* data) {
+                             TfLiteDepthwiseConvParams* params, int width,
+                             int height, int filter_width, int filter_height,
+                             const TfLiteType data_type, OpData* data) {
   bool has_bias = node->inputs->size == 3;
   // Check number of inputs/outputs
   TF_LITE_ENSURE(context, has_bias || node->inputs->size == 2);
   TF_LITE_ENSURE_EQ(context, node->outputs->size, 1);
 
-  // Matching GetWindowedOutputSize in TensorFlow.
-  auto padding = params->padding;
+  int unused_output_height, unused_output_width;
   data->padding = ComputePaddingHeightWidth(
-      params->stride_height, params->stride_width,
-      params->dilation_height_factor, params->dilation_width_factor, height,
-      width, filter_height, filter_width, padding, &out_height, &out_width);
+      params->stride_height, params->stride_width, 1, 1, height, width,
+      filter_height, filter_width, params->padding, &unused_output_height,
+      &unused_output_width);
 
   // Note that quantized inference requires that all tensors have their
   // parameters set. This is usually done during quantized training.
@@ -82,15 +80,13 @@ TfLiteStatus CalculateOpData(TfLiteContext* context, TfLiteNode* node,
     const TfLiteTensor* bias =
         GetOptionalInputTensor(context, node, kBiasTensor);
     TfLiteTensor* output = GetOutput(context, node, kOutputTensor);
-    int output_channels = filter->dims->data[kConvQuantizedDimension];
-
-    TF_LITE_ENSURE_STATUS(tflite::PopulateConvolutionQuantizationParams(
+    int num_channels = filter->dims->data[kDepthwiseConvQuantizedDimension];
+    return tflite::PopulateConvolutionQuantizationParams(
         context, input, filter, bias, output, params->activation,
         &data->output_multiplier, &data->output_shift,
         &data->output_activation_min, &data->output_activation_max,
         data->per_channel_output_multiplier,
-        reinterpret_cast<int*>(data->per_channel_output_shift),
-        output_channels));
+        reinterpret_cast<int*>(data->per_channel_output_shift), num_channels);
   }
   return kTfLiteOk;
 }
@@ -100,12 +96,12 @@ TfLiteStatus CalculateOpData(TfLiteContext* context, TfLiteNode* node,
  * calculated only once, not in every invocation.
  */
 template <typename T>
-inline void PrecomputeSumOfFilters(const int32_t* bias,
-                                   const TfLiteTensor* filters,
-                                   int32_t* sum_of_filters_factor,
-                                   RuntimeShape filter_shape,
-                                   int32_t input_offset,
-                                   int32_t filter_offset = 0) {
+inline void PrecomputeSumOfFiltersFactor(const int32_t* bias,
+                                         const TfLiteTensor* filters,
+                                         int32_t* sum_of_filters_factor,
+                                         RuntimeShape filter_shape,
+                                         int32_t input_offset,
+                                         int32_t filter_offset = 0) {
   if (filters->type == kTfLiteInt8) {
     // Ensure that the filter offset is 0 in the signed integer case
     TFLITE_DCHECK_EQ(filter_offset, 0);
@@ -113,36 +109,22 @@ inline void PrecomputeSumOfFilters(const int32_t* bias,
   const T* filter_data = GetTensorData<T>(filters);
   const int filter_height = filter_shape.Dims(1);
   const int filter_width = filter_shape.Dims(2);
-  const int input_depth = filter_shape.Dims(3);
-  const int output_depth = filter_shape.Dims(0);
+  const int num_filters = filter_shape.Dims(3);
 
-  int filter_size = filter_width * filter_height * input_depth;
+  int filter_size = filter_width * filter_height;
 
-  for (int out_channel = 0; out_channel < output_depth; ++out_channel) {
+  for (int out_channel = 0; out_channel < num_filters; ++out_channel) {
     int32_t sum_of_filter_factor = filter_size * filter_offset;
 
-    for (int filter_index = filter_size * out_channel;
-         filter_index < filter_size + filter_size * out_channel;
-         ++filter_index) {
+    for (int filter_index = out_channel;
+         filter_index < filter_size * num_filters;
+         filter_index += num_filters) {
       sum_of_filter_factor += filter_data[filter_index];
     }
     sum_of_filters_factor[out_channel] = sum_of_filter_factor * input_offset;
 
     if (bias) {
       sum_of_filters_factor[out_channel] += bias[out_channel];
-    }
-  }
-};
-
-inline void PrecomputeSumOfFiltersPadding(const int32_t* bias,
-                                          int32_t* sum_of_filters_factor,
-                                          RuntimeShape filter_shape) {
-  const int output_depth = filter_shape.Dims(0);
-  for (int out_channel = 0; out_channel < output_depth; ++out_channel) {
-    if (bias) {
-      sum_of_filters_factor[out_channel] = bias[out_channel];
-    } else {
-      sum_of_filters_factor[out_channel] = 0;
     }
   }
 }
@@ -153,9 +135,12 @@ void Free(TfLiteContext* context, void* buffer) {}
  * Evaluation function. Called in every invocation.
  */
 TfLiteStatus Eval(TfLiteContext* context, TfLiteNode* node) {
+  TfLiteDepthwiseConvParams& params =
+      *(reinterpret_cast<TfLiteDepthwiseConvParams*>(node->builtin_data));
   OpData* data = reinterpret_cast<OpData*>(node->user_data);
-  auto* params = reinterpret_cast<TfLiteConvParams*>(node->builtin_data);
 
+  TfLiteEvalTensor* output =
+      tflite::micro::GetEvalOutput(context, node, kOutputTensor);
   const TfLiteEvalTensor* input =
       tflite::micro::GetEvalInput(context, node, kInputTensor);
   const TfLiteEvalTensor* filter =
@@ -164,16 +149,15 @@ TfLiteStatus Eval(TfLiteContext* context, TfLiteNode* node) {
       (NumInputs(node) == 3)
           ? tflite::micro::GetEvalInput(context, node, kBiasTensor)
           : nullptr;
-  TfLiteEvalTensor* output =
-      tflite::micro::GetEvalOutput(context, node, kOutputTensor);
 
-  return data->eval_function(params, data, input, filter, bias, output,
-                             context);
+  return data->eval_function(context, params, data, input, filter, bias,
+                             output);
 }
 
-}  // namespace conv
+}  // namespace depthwise_conv
 }  // namespace micro
 }  // namespace ops
 }  // namespace tflite
 
-#endif /* TENSORFLOW_LITE_MICRO_KERNELS_IFX_FAST_CONV_CONV_IMPL_H_ */
+#endif /* TENSORFLOW_LITE_MICRO_KERNELS_GENERIC_FAST_DEPTHWISE_CONV_DEPTHWISE_CONV_IMPL_H_ \
+        */
