+++ tensorflow/lite/micro/kernels/generic_fast/reduce/reduce_impl.h
@@ -13,8 +13,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-#ifndef TENSORFLOW_LITE_MICRO_KERNELS_IFX_FAST_REDUCE_REDUCE_IMPL_H_
-#define TENSORFLOW_LITE_MICRO_KERNELS_IFX_FAST_REDUCE_REDUCE_IMPL_H_
+#ifndef TENSORFLOW_LITE_MICRO_KERNELS_GENERIC_FAST_REDUCE_REDUCE_IMPL_H_
+#define TENSORFLOW_LITE_MICRO_KERNELS_GENERIC_FAST_REDUCE_REDUCE_IMPL_H_
 
 #include "tensorflow/lite/c/builtin_op_data.h"
 #include "tensorflow/lite/c/common.h"
@@ -23,7 +23,7 @@ limitations under the License.
 #include "tensorflow/lite/kernels/internal/tensor_ctypes.h"
 #include "tensorflow/lite/kernels/internal/types.h"
 #include "tensorflow/lite/kernels/kernel_util.h"
-#include "tensorflow/lite/micro/kernels/ifx_fast/reduce/reduce_op_data.h"
+#include "tensorflow/lite/micro/kernels/generic_fast/reduce/reduce_op_data.h"
 #include "tensorflow/lite/micro/kernels/kernel_util.h"
 #include "tensorflow/lite/micro/micro_utils.h"
 
@@ -752,4 +752,4 @@ TfLiteStatus EvalMeanMax(TfLiteContext* context, TfLiteNode* node) {
 }  // namespace ops
 }  // namespace tflite
 
-#endif /* TENSORFLOW_LITE_MICRO_KERNELS_IFX_FAST_REDUCE_REDUCE_IMPL_H_ */
+#endif // TENSORFLOW_LITE_MICRO_KERNELS_GENERIC_FAST_REDUCE_REDUCE_IMPL_H_
