+++ tensorflow/lite/micro/memory_arena_threshold_test.cc
@@ -49,7 +49,7 @@ constexpr int kKeywordModelNodeAndRegistrationCount = 15;
 //
 // Run this test with '--copt=-DTF_LITE_STATIC_MEMORY' to get optimized memory
 // runtime values:
-#ifdef IFX_FAST
+#ifdef GENERIC_FAST
 constexpr int kKeywordModelTotalSize = 14736;
 constexpr int kKeywordModelTailSize = 14064;
 constexpr int kKeywordModelPersistentTfLiteTensorDataSize = 128;
@@ -78,7 +78,7 @@ constexpr int kTestConvModelNodeAndRegistrationCount = 7;
 
 // NOTE: These values are measured on x86-64:
 // TODO(b/158651472): Consider auditing these values on non-64 bit systems.
-#ifdef IFX_FAST
+#ifdef GENERIC_FAST
 constexpr int kTestConvModelTotalSize = 10048;
 constexpr int kTestConvModelTailSize = 2304;
 constexpr int kTestConvModelPersistentTfLiteTensorDataSize = 128;
